// frontend/src/app/domain/models/entities/user-register.entitie.ts
export interface UserUpdate {
    arlInsurance: string;
    birthdate: string;
    cesantias: string;
    company: string;
    email: string;
    healthyInsurance: string;
    identifier: string;
    lastName: string;
    name: string;
    pension: string;
    position: string;
    privileges: string;
    salary: string;
    username: string;
    password: string;
    transactUser: string;
    account: string;
}